+++ Templates/DTL-SharePoint-FullConfig/dsc/ConfigureSPVM.ps1
@@ -18,7 +18,7 @@ configuration ConfigureSPVM
         [Parameter(Mandatory)] [System.Management.Automation.PSCredential]$SPSuperReaderCreds
     )
 
-    Import-DscResource -ModuleName ComputerManagementDsc, NetworkingDsc, ActiveDirectoryDsc, xCredSSP, xWebAdministration, SharePointDsc, xPSDesiredStateConfiguration, xDnsServer, CertificateDsc, SqlServerDsc
+    Import-DscResource -ModuleName ComputerManagementDsc, NetworkingDsc, ActiveDirectoryDsc, xCredSSP, xWebAdministration, SharePointDsc, xPSDesiredStateConfiguration, xDnsServer, CertificateDsc, SqlServerDsc, cChoco
 
     [String] $DomainNetbiosName = (Get-NetBIOSName -DomainFQDN $DomainFQDN)
     $Interface = Get-NetAdapter| Where-Object Name -Like "Ethernet*"| Select-Object -First 1
@@ -184,6 +184,62 @@ configuration ConfigureSPVM
                 Set-ItemProperty -Path $ieNewTabKey -Name "NewTabPageShow" -Value 0
             }
             GetScript = { }
+        }
+
+        #**********************************************************
+        # Install applications using Chocolatey
+        #**********************************************************
+        cChocoInstaller InstallChoco
+        {
+            InstallDir = "C:\Program Files\Choco"
+        }
+
+        cChocoPackageInstaller InstallEdge
+        {
+            Name                 = 'microsoft-edge'
+            Ensure               = 'Present'
+            Version              =  83.0.478.61
+            DependsOn            = '[cChocoInstaller]InstallChoco'
+        }
+
+        cChocoPackageInstaller InstallChrome
+        {
+            Name                 = 'googlechrome'
+            Ensure               = 'Present'
+            Version              =  83.0.4103.116
+            DependsOn            = '[cChocoInstaller]InstallChoco'
+        }
+
+        cChocoPackageInstaller InstallEverything
+        {
+            Name                 = 'everything'
+            Ensure               = 'Present'
+            Version              =  1.4.1969
+            DependsOn            = '[cChocoInstaller]InstallChoco'
+        }
+
+        cChocoPackageInstaller InstallILSpy
+        {
+            Name                 = 'ilspy'
+            Ensure               = 'Present'
+            Version              =  6.0.0.5836
+            DependsOn            = '[cChocoInstaller]InstallChoco'
+        }
+
+        cChocoPackageInstaller InstallNotepadpp
+        {
+            Name                 = 'notepadplusplus.install'
+            Ensure               = 'Present'
+            Version              =  7.9
+            DependsOn            = '[cChocoInstaller]InstallChoco'
+        }
+
+        cChocoPackageInstaller Install7zip
+        {
+            Name                 = '7zip.install'
+            Ensure               = 'Present'
+            Version              =  19.0
+            DependsOn            = '[cChocoInstaller]InstallChoco'
         }
 
         #**********************************************************
@@ -424,6 +480,15 @@ configuration ConfigureSPVM
             PsDscRunAsCredential = $SPSetupCredential
             DependsOn            = "[Group]AddSPSetupAccountToAdminGroup", "[ADUser]CreateSParmAccount", "[ADUser]CreateSPSvcAccount", "[ADUser]CreateSPAppPoolAccount", "[ADUser]CreateSPSuperUserAccount", "[ADUser]CreateSPSuperReaderAccount", "[xScript]CreateWSManSPNsIfNeeded"
         }
+        
+        cChocoPackageInstaller InstallFiddler
+        {
+            Name                 = 'fiddler'
+            Ensure               = 'Present'
+            Version              =  5.0.20202.18177
+            PsDscRunAsCredential = $DomainAdminCredsQualified
+            DependsOn            = '[cChocoInstaller]InstallChoco'
+        }
 
         xScript WaitForSQL
         {
@@ -639,14 +704,14 @@ configuration ConfigureSPVM
             {
                 # If the TestScript returns $false, DSC executes the SetScript to bring the node back to the desired state
                 TestScript = {
-                    return (Test-Path HKLM:\SOFTWARE\SPDSCConfigForceRebootKey\RebootRequested)
+                    return (Test-Path HKLM:\SOFTWARE\DscScriptExecution\flag_ForceRebootBeforeCreatingSPTrust)
                 }
                 SetScript = {
-                    New-Item -Path HKLM:\SOFTWARE\SPDSCConfigForceRebootKey\RebootRequested -Force
+                    New-Item -Path HKLM:\SOFTWARE\DscScriptExecution\flag_ForceRebootBeforeCreatingSPTrust -Force
                     $global:DSCMachineStatus = 1
                 }
                 GetScript = { }
-                PsDscRunAsCredential = $SPSetupCredsQualified
+                PsDscRunAsCredential = $DomainAdminCredsQualified
                 DependsOn = "[SPFarmSolution]InstallLdapcp"
             }