+++ Templates/SharePoint-ADFS/CHANGELOG.md
@@ -4,9 +4,16 @@
 
 * Set FrontEnd VMs with SharePoint MinRole Front-End on SharePoint versions that support MinRoles configuration
 * Increase max numberOfAdditionalFrontEnd from 3 to 4
+* Install Edge in SharePoint VM through Chocolatey
+* Install Chrome in SharePoint VM through Chocolatey
+* Install Fiddler in SharePoint VM through Chocolatey
+* Install Notepad++ in SharePoint VM through Chocolatey
+* Install 7-zip in SharePoint VM through Chocolatey
+* Install Everything in SharePoint VM through Chocolatey
 * Define the list of all possible values for the time zone parameter vmsTimeZone
 * Update WaitToAvoidServersJoiningFarmSimultaneously to ensure it runs only 1 time, and updated the delay from 60 to 90 secs to improve reliability
 * Use a unique location for custom registry keys
+* Update parameters passed to ConfigureFE
 * Update DSC module SharePointDSC from 4.2 to 4.3
 * Update DSC module NetworkingDsc from 8.0 to 8.1
 * Update DSC module ActiveDirectoryCSDsc from 4.1 to 5.0